EFS-283 Fixed test cases
