+++ src/admin/utils/fixMultipleSourceAssigningAuthority.js
@@ -12,6 +12,10 @@ const { SecurityTagSystem } = require('../../utils/securityTagSystem');
 
 function fixPractitionerResource(resource, fixMultipleOwners) {
     let security = resource.meta.security || [];
+    if (!security.length){
+        console.log(`meta.security not present for resource _id: ${resource._id}. Skipping`);
+        return resource;
+    }
     const sourceAssigningAuthorities = security.filter(s => s.system === SecurityTagSystem.sourceAssigningAuthority);
     const source = resource.meta.source;
     let securityWithoutOriginalOwnersAndAuthority;
@@ -65,6 +69,10 @@ function fixPractitionerResource(resource, fixMultipleOwners) {
 
 function fixResource(resource) {
     let security = resource.meta.security || [];
+    if (!security.length){
+        console.log(`meta.security not present for resource _id: ${resource._id}. Skipping`);
+        return resource;
+    }
     const sourceAssigningAuthorities = security.filter(s => s.system === SecurityTagSystem.sourceAssigningAuthority);
     const owners = security.filter(s => s.system === SecurityTagSystem.owner);
     let securityWithoutOriginalAuthority = security.filter(
